additional cleaning up after moving primitives folder
